UDP Exporter release workflow (#1049)
*Description of changes:*
The validation steps for the GitHub workflow intended to manually
trigger a UDP exporter release to Maven Central includes:
The workflow will first build the UDP exporter.
The workflow will then set up and run an instance of X-Ray daemon.
Following setup, it will run a validation app to send traces to X-Ray
daemon via UDP exporter.
Finally, it will execute a validation step to verify that daemon
received the traces.

TODO:
1. Add a publish to Maven Central step to release workflow.

Test workflow link:
https://github.com/aws-observability/aws-otel-java-instrumentation/actions/runs/14043198263

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------
