--- exporters/aws-opentelemetry-xray-lambda-exporter/settings.gradle.kts
@@ -1,12 +1,0 @@
-rootProject.name = "aws-opentelemetry-xray-lambda-exporter"
-
-dependencyResolutionManagement {
-  repositories {
-    mavenCentral()
-    mavenLocal()
-
-    maven {
-      setUrl("https://oss.sonatype.org/content/repositories/snapshots")
-    }
-  }
-}
