+++ exporters/aws-opentelemetry-xray-udp-span-exporter/settings.gradle.kts
@@ -1,0 +1,12 @@
+rootProject.name = "aws-opentelemetry-xray-udp-span-exporter"
+
+dependencyResolutionManagement {
+  repositories {
+    mavenCentral()
+    mavenLocal()
+
+    maven {
+      setUrl("https://oss.sonatype.org/content/repositories/snapshots")
+    }
+  }
+}
