+++ sample-apps/integ-test-app/settings.gradle.kts
@@ -1,0 +1,12 @@
+rootProject.name = "integ-test-app"
+
+dependencyResolutionManagement {
+  repositories {
+    mavenCentral()
+    mavenLocal()
+
+    maven {
+      setUrl("https://oss.sonatype.org/content/repositories/snapshots")
+    }
+  }
+}
